chore: move PhelFnNormalizer to Application layer
