+++ src/php/Api/ApiFactory.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Phel\Api;
 
 use Gacela\Framework\AbstractFactory;
-use Phel\Api\Domain\PhelFnNormalizer;
+use Phel\Api\Application\PhelFnNormalizer;
 use Phel\Api\Domain\PhelFnNormalizerInterface;
 use Phel\Api\Infrastructure\PhelFnLoader;
 use Phel\Api\Infrastructure\PhelFnLoaderInterface;
